requirements.txt
django
psycopg2
djangorestframework
django-filter
dj-rest-auth